requirements.txt
openai
streamlit
requests
geocoder
googlemaps
python-dotenv
pillow